feat(matcher): E1.2 — move UserToTrialAttrMatcher into exact_matching
Relocate the 1137-line per-trial matcher into the package
(exact_matching/matcher.py); trials/services/user_to_trial_attr_matcher.py
becomes a back-compat shim re-exporting the public API AND the private
_resolve_omop_concepts (imported by tests) so the move stays transparent for all
importers (models, queryset, serializers, mapper, commands, tests). The matcher's
therapy_match_profile imports (top-level + lazy) now point at the package.

Its other top-level deps (configs, patient_info_attributes, genetic_mutations,
flipi, utils, enums) still import from  for now — the package therefore
depends on EXACT's trials app at import time; full CB-self-containment (moving
those pure-logic deps) is E1.3. The ~6 model-access sites (vocab-mirror titles,
therapy component/category expansion) are decoupled behind a data-access port in
E1.2b.

Full EXACT suite green (1073 passed, 5 skipped).

Part of EPIC cancerbot#4601 (E1 #272). Proposal — cross-team.
